+++ include/NeoN/dsl/solver.hpp
@@ -37,13 +37,14 @@ la::SolverStats iterativeSolveImpl(
     std::vector<const PostAssemblyBase<typename VectorType::ElementType, IndexType>*> ps = {}
 )
 {
-    exp.read(fvSchemes);
-    exp.assemble(t, dt, ls, ps);
+    auto optExp = optimize(exp);
+    optExp.read(fvSchemes);
+    optExp.assemble(t, dt, ls, ps);
 
     // TODO move that to expression explicit operation or
     // into functor ?
     // subtract the explicit source term from the rhs
-    auto expTmp = exp.explicitOperation(solution.mesh().nCells());
+    auto expTmp = optExp.explicitOperation(solution.mesh().nCells());
     auto [vol, expSource, rhs] = views(solution.mesh().cellVolumes(), expTmp, ls.rhs());
     parallelFor(
         solution.exec(),